:recycle: refactor: Evaluator 기능 분리 및 util 테스트 추가
