+++ src/main/java/util/evaluator/TimeEvaluator.java
@@ -1,0 +1,13 @@
+package util.evaluator;
+
+import java.time.LocalTime;
+
+public class TimeEvaluator {
+
+    private TimeEvaluator() {
+    }
+
+    public static boolean isOpenTime(LocalTime time) {
+        return time.isAfter(LocalTime.of(8, 0)) && time.isBefore(LocalTime.of(23, 0));
+    }
+}
