+++ src/test/java/util/reader/CsvReaderTest.java
@@ -1,4 +1,4 @@
-package util;
+package util.reader;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
